Feat: continuing work on merge
- Fixing compilation errors between Sova and doux-sova
- Introducing a CHANGELOG.md to better track changes to the repo

+++ core/src/clock.rs
@@ -303,7 +303,7 @@ impl Clock {
     /// Start/stop synchronization feature in Ableton Link.
     pub fn play_pause(&mut self) {
         self.session_state
-            .set_is_playing(!self.is_playing(), self.micros());
+            .set_is_playing(!self.is_playing(), self.micros() as i64);
         self.commit_app_state();
     }
